backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
fpdf2>=2.7.0
duckduckgo-search>=6.0.0
python-multipart>=0.0.9
openai>=1.0.0   # optional — AI summarization
